refactor(nms): Separate the TreaceContext
